- use double precision to store the position / zoom / other stuff - some temporary fix to get a +50% improvement from vectorization until   we have vectorisation for comparisons and redux

+++ demos/mandelbrot/mandelbrot.h
@@ -9,8 +9,8 @@ class MandelbrotWidget : public QWidget
 {
     Q_OBJECT
 
-    Eigen::Vector2f center;
-    float xradius;
+    Eigen::Vector2d center;
+    double xradius;
     int size;
     unsigned char *buffer;
     QPoint lastpos;
